+++ packages/sdk/src/index.ts
@@ -300,6 +300,7 @@ export {
   getChainIdByName,
   getChainNameById,
   getEthNetworkByEnv,
+  getLegacyChainNameById,
   getSolanaNetworkByEnv,
   getSonicNetworkByEnv,
   getStarknetNetworkByEnv,
@@ -525,6 +526,12 @@ export {
   type IGetUserStakeAndBakeSignatureParams,
   type IGetUserStakeAndBakeSignatureResponse,
 } from './api-functions/getUserStakeAndBakeSignature/getUserStakeAndBakeSignature';
+export {
+  canResolveDepositBtcAddressWithJwt,
+  getDepositAssetTypeById,
+  type IResolveDepositBtcAddressParams,
+  resolveDepositBtcAddress,
+} from './api-functions/resolveDepositBtcAddress/resolveDepositBtcAddress';
 export {
   type ISetReferralParams,
   setReferral,